TEST: test all behaviors with both cholmod and SuperLU
